% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark data}
\format{A \code{list} with a \code{data.frame} for each model family (Gaussian, binomial,
multinomial), featuring 5 variables:
\describe{
\item{dataset}{dataset used}
\item{penalty}{type of penalty (ridge or lasso)}
\item{package}{R package}
\item{time}{run time in seconds}
\item{loss}{objective loss}
}}
\source{
<https://github.com/jolars/sgdnet/data-raw/
}
\usage{
benchmarks
}
\description{
Benchmark data
}
\keyword{datasets}
