% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benchmarks_binomial}
\alias{benchmarks_binomial}
\title{Benchmark data for binomial response family}
\format{A \code{data.frame} with 5 variables and 2000 observations.
\describe{
\item{dataset}{dataset used}
\item{package}{R package}
\item{time}{run time in seconds}
\item{loss}{objective loss}
\item{penalty}{type of penalty, "ridge" or "lasso"}
}}
\source{
<https://github.com/jolars/sgdnet/data-raw/
}
\usage{
benchmarks_binomial
}
\description{
Benchmark data for binomial response family
}
\keyword{datasets}
