% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.sgdnet.R
\name{bind_intercept}
\alias{bind_intercept}
\title{Bind intercept with coefficients}
\usage{
bind_intercept(beta, a0)
}
\arguments{
\item{beta}{coefficients}

\item{a0}{intercept}
}
\value{
A matrix (or listof matrices).
}
\description{
Bind intercept with coefficients
}
\keyword{internal}
