% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{caddisfly}
\alias{caddisfly}
\title{Caddisfly}
\format{49 observations from 12 variables represented as a list consisting
of a numeric response vector \code{y} indicating counts of caddisfly (Stactobiella risi)
meteorological factor as well as \code{x}: a dense feature matrix
with the following variables:
\describe{
\item{t_max}{maximum temperature of day j}
\item{t_soir}{average of temperature at 18h and 21h}
\item{t_min}{minimum temperature of day (j+1)}
\item{vent}{wind speed}
\item{pression}{average pressure of 0h, 6h, 12h and 18h}
\item{var_pression}{difference between day j and day (j+1)'s pressure at 6h }
\item{humidite}{humidity at 21h}
\item{nebu_nuit}{average value of cloudiness at 21h}
\item{precip_nuit}{precipitation of the night}
\item{nebu_moy}{average value of cloudiness at day j's 9h, 12h, 15h, 18h, 21h
and day (j+1)'s 0h, 3h, 6h}
\item{precip_tot}{total precipitation at 6h of day j and day (j+1)}
\item{cla}{night id}
}}
\source{
Usseglio-Polatera, P., and Y. Auda. (1987).
Influence Des Facteurs Météorologiques Sur Les Résultats de Piégeage Lumineux,
In Annales de Limnologie-International Journal of Limnology, 23:65–79. 1. EDP Sciences.

\url{http://pbil.univ-lyon1.fr/R/pdf/pps034.pdf}

\url{https://pbil.univ-lyon1.fr/ADE-4/ade4-html/trichometeo.html}
}
\usage{
caddisfly
}
\description{
This data set describes abundances of caddisfly and meteorological features
that may influence its presence during the 49 trapping night. The original dataset
contains counts for 17 species, this dataset use one of them, which is Stactobiella risi
}
\keyword{datasets}
