% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.sgdnet.R
\name{coef.sgdnet}
\alias{coef.sgdnet}
\title{Extract Model Coefficients for sgdnet Model}
\usage{
\method{coef}{sgdnet}(object, ...)
}
\arguments{
\item{object}{a model of class \code{'sgdnet'}, typically from a call to
\code{\link[=sgdnet]{sgdnet()}}.}

\item{...}{passed on to \code{\link[=predict.sgdnet]{predict.sgdnet()}}}
}
\value{
A sparse matrix with intercept in the first row and betas
in the rest.
}
\description{
This is simply a wrapper for \code{\link[=predict.sgdnet]{predict.sgdnet()}} with \code{type = "coefficients"}.
}
\examples{
fit <- sgdnet(matrix(rnorm(100), 50, 2), rnorm(50))
coef(fit)
}
\seealso{
\code{\link[=predict.sgdnet]{predict.sgdnet()}}, \code{\link[=sgdnet]{sgdnet()}}
}
