% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_family}
\alias{extract_family}
\title{Retrieve Response Family from sgdnet Object}
\usage{
extract_family(x)
}
\arguments{
\item{x}{an object of class \code{'sgdnet'}}
}
\value{
The family that was used in fitting \code{x}.
}
\description{
Retrieve Response Family from sgdnet Object
}
\keyword{internal}
