% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FitModel}
\alias{FitModel}
\title{Fit a Model with sgdnet}
\usage{
FitModel(x_in, y, family_in, fit_intercept, is_sparse, alpha, beta, normalize,
  max_iter, tol, debug)
}
\arguments{
\item{y}{response matrix}

\item{family_in}{the response type}

\item{is_sparse}{is x sparse?}

\item{alpha}{l2-regularization penalty}

\item{beta}{l1-regularization penalty}

\item{normalize}{should x be normalized before fitting the model?}

\item{max_iter}{the maximum number of iterations}

\item{tol}{tolerance for convergence. The algorithm terminates when
max(change in weights)/max(weights) < tol.}

\item{x}{feature matrix}
}
\value{
A list of
\itemize{
\item ao: the intercept
\item beta: the weights
\item losses: the loss at each outer iteration
\item nseen: the number of samples seen
\item npasses: the number of effective passes (epochs)
\item return_code: the convergence result. 0 mean that the algorithm converged,
1 means that \code{max_iter} was reached before the algorithm converged.
}
}
\description{
Fit a Model with sgdnet
}
\keyword{internal}
