% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{houses}
\alias{houses}
\title{S&P Letters Data}
\format{A list with two items representing 20,640 observations from
9 variables
\describe{
\item{x}{a dense feature matrix with median house value, median income,
housing median age, total rooms, total bedrooms, population,
households, latitude, and longitude}
\item{y}{a numeric vector of permeability values for 165 compounds}
}}
\source{
Pace, R. Kelley and Ronald Barry, Sparse Spatial Autoregressions,
Statistics and Probability Letters, 33 (1997) 291-297.
}
\usage{
houses
}
\description{
Spatial data containing housing prices with economic covariates.
}
\details{
"We collected information on the variables using all the block groups in
California from the 1990 Census. In this sample a block group on average
includes 1425.5 individuals living in a geographically compact area.
Naturally, the geographical area included varies inversely with the
population density. We computed distances among the centroids of each
block group as measured in latitude and longitude. We excluded all the block
groups reporting zero entries for the independent and dependent variables.
The final data contained 20,640 observations on 9 variables. The dependent
variable is ln(median house value)."
}
\keyword{datasets}
