% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.sgdnet.R
\name{interpolate_coefficients}
\alias{interpolate_coefficients}
\title{Interpolate coefficients}
\usage{
interpolate_coefficients(beta, s, lamlist)
}
\arguments{
\item{beta}{coeffieicnts}

\item{s}{lambda penalty}

\item{lamlist}{a list of lambda interpolation parameters, returned from
\code{\link[=lambda_interpolate]{lambda_interpolate()}}.}
}
\value{
A matrix (or list of matrices) with new coefficients based
on linearly interpolating from new and old lambda values.
}
\description{
Interpolate coefficients
}
\keyword{internal}
