% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.sgdnet.R
\name{lambda_interpolate}
\alias{lambda_interpolate}
\title{Interpolate Lambda Values}
\usage{
lambda_interpolate(lambda, s)
}
\arguments{
\item{lambda}{lambda penalty}

\item{s}{the lambda penalty wanted by the user}
}
\value{
Interpolated values of lambda
}
\description{
Interpolate
}
\author{
Jerome Friedman, Trevor Hastie, Rob Tibshirani, and Noah Simon
}
\keyword{internal}
