% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mushrooms}
\alias{mushrooms}
\title{Mushrooms}
\format{A list with two items representing 8,124 observations from
112 variables.
\describe{
\item{x}{a sparse feature matrix dummy-coded attributes for each
mushroom}
\item{y}{a factor variable represing the outcome, edibility, either
"edible" or "poisonous"}
}}
\source{
The Audubon Society Field Guide to North
American Mushrooms (1981). G. H. Lincoff (Pres.), New York: Alfred A.
Knopf.

\url{https://archive.ics.uci.edu/ml/datasets/mushroom}

Dua, D. and Karra Taniskidou, E. (2017).
UCI Machine Learning Repository \link{http://archive.ics.uci.edu/ml}. Irvine,
CA: University of California, School of Information and Computer Science.
}
\usage{
mushrooms
}
\description{
Mushroom records from The Audubon Society Field Guide to North
American Mushrooms (1981). G. H. Lincoff (Pres.), New York: Alfred A. Knopf.
}
\details{
The data set was retrieved from the UCI database and came with the
following description:

"This data set includes descriptions of hypothetical samples corresponding to
23 species of gilled mushrooms in the Agaricus and Lepiota Family.
Each species is identified as definitely edible, definitely
poisonous, or of unknown edibility and not recommended. This latter class was
combined with the poisonous one. The Guide clearly states that there is no
simple rule for determining the edibility of a mushroom; no rule like
"leaflets three, let it be" for Poisonous Oak and Ivy."
}
\section{Processing}{

The original dataset contained 22 variables (features). Out of these,
"stalk-root" was dropped since it contained missing values. All of the
remaining variables were dummy-coded.
}

\keyword{datasets}
