% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.sgdnet.R
\name{nonzero_coefs}
\alias{nonzero_coefs}
\title{Return Non-zero Coefficients}
\usage{
nonzero_coefs(beta, bystep = FALSE)
}
\arguments{
\item{beta}{ceofficients}

\item{bystep}{\code{FALSE} selects variables that were ever nonzero.
\code{TRUE} selects which variables are nonzero at each step along the
path.}
}
\value{
Nonzero coefficients.
}
\description{
Return Non-zero Coefficients
}
\author{
Jerome Friedman, Trevor Hastie, Rob Tibshirani, and Noah Simon
}
\keyword{internal}
