% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{permeability}
\alias{permeability}
\title{Permeability}
\format{A list with two items representing 165 observations from
1107 variables.
\describe{
\item{x}{a sparse feature matrix with binary indicators for molecular
fingerprints labeled X1 to X1107}
\item{y}{a numeric vector of permeability values for 165 compounds}
}}
\source{
\link[AppliedPredictiveModeling:permeability]{AppliedPredictiveModeling::permeability}, which features a
thorough description on the type of data.
}
\usage{
permeability
}
\description{
A pharmaceutical dataset of permeability values for 165 compounds
for which molecular fingerprints were collected and represented
as binary indices.
}
\keyword{datasets}
