% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.sgdnet.R
\name{predict.sgdnet}
\alias{predict.sgdnet}
\title{Predictions for sgdnet Models}
\usage{
\method{predict}{sgdnet}(object, newx, s = NULL, type = c("link",
  "response", "coefficients", "nonzero", "class"), exact = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{'sgdnet'}.}

\item{newx}{new data to predict on. Must be provided if \code{type} is
\code{"link"}.}

\item{s}{the lambda penalty value on which to base the predictions.}

\item{type}{type of prediction to return, one of
\describe{
\item{\code{link}}{ linear predictors,}
\item{\code{response}}{responses,}
\item{\code{coefficients}}{coefficients (weights); equivalent to calling
\code{\link[=coef]{coef()}}, and}
\item{\code{nonzero}}{nonzero coefficients at each step of the regularization
path.}
}}

\item{exact}{if the given value of \code{s} is not in the model and
\code{exact = TRUE}, the model will be refit using \code{s}. If \code{FALSE}, predictions
will be made using a linearly interpolated coefficient matrix.}

\item{...}{arguments to be passed on to \code{\link[stats:update]{stats::update()}} to refit
the model via \code{\link[=sgdnet]{sgdnet()}} if \code{s} is missing
from the model and an exact fit is required by \code{exact}.}
}
\value{
Predictions for \code{object} given data in \code{newx}.
}
\description{
Predictions for sgdnet Models
}
\examples{
# Gaussian

# Split into training and test sets
id <- sample.int(nrow(iris))
train_ind <- id[1:100]
test_ind <- id[101:150]
gaussian_fit <- sgdnet(iris[train_ind, 2:4], iris[train_ind, 1])
pred <- predict(gaussian_fit,
                newx = iris[test_ind, 2:4],
                s = 0.003,
                type = "response",
                exact = TRUE)

# Mean absolute error
mae <- 1/length(test_ind)*sum(abs(iris[test_ind, 2:4] - pred))
}
\seealso{
\code{\link[=sgdnet]{sgdnet()}}, \code{\link[=coef.sgdnet]{coef.sgdnet()}}
}
\author{
Jerome Friedman, Trevor Hastie, Rob Tibshirani, Noah Simon
(original), Johan Larsson (modifications)
}
