% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cv_sgdnet.R
\name{print.cv_sgdnet}
\alias{print.cv_sgdnet}
\title{Print Method for \code{'cv_sgdnet'} objects}
\usage{
\method{print}{cv_sgdnet}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"cv_sgdnet"} as generated from a call to
\code{\link[=cv_sgdnet]{cv_sgdnet()}}}

\item{...}{arguments passed on to \code{\link[=print]{print()}}}
}
\value{
Prints a \link{data.frame} with columns
\item{\code{type}}{type of validation performance metric}
\item{\code{alpha}}{elastic net mixing parameter}
\item{\code{lambda}}{regularization strength}
\item{\code{mean}}{mean of the prediction metric across the folds}
\item{\code{sd}}{standard deviation of the prediction metric across the folds}
\item{\code{ci_lo}}{mean minus one standard deviation}
\item{\code{ci_up}}{mean plus one standard deviation}
}
\description{
This function prints the cross-validation summary for the \code{alpha} with
the best validation performance as well as the \code{lambda} at which this
performance was reached and the largest \code{lambda} (\code{lambda_1se}) with a
performance within one standard deviation of that.
}
\examples{
fit <- cv_sgdnet(mtcars$drat, mtcars$hp)
print(fit)
}
\seealso{
\code{\link[=print.data.frame]{print.data.frame()}}, \code{\link[=print]{print()}}, \code{\link[=cv_sgdnet]{cv_sgdnet()}}
}
