% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sgdnet.R
\name{print.sgdnet}
\alias{print.sgdnet}
\title{Print a Path Summary for a Fit from sgdnet}
\usage{
\method{print}{sgdnet}(x, ...)
}
\arguments{
\item{x}{an object of class \code{'sgdnet'}, typically the result of
calling \code{\link[=sgdnet]{sgdnet()}}}

\item{...}{other arguments passed to \code{\link[=print]{print()}}}

\item{digits}{the number of significant digits to print}
}
\value{
Prints (and return invisibly) a table of the regularization path from
the fit with the following columns:
\tabular{ll}{
\code{Df}     \tab pseudo-degrees of freedom, namely the number of non-zero
coefficients \cr
\code{\%Dev}   \tab the percentage of deviance explained \cr
\code{Lambda} \tab the \eqn{\lambda}{lambda} value (regularization strength) of
the fit
}
}
\description{
Print a Path Summary for a Fit from sgdnet
}
\examples{
fit <- sgdnet(with(mtcars, cbind(drat, hp)), mtcars$disp)
print(fit, digits = 1)
}
\seealso{
\code{\link[=sgdnet]{sgdnet()}}, \code{\link[=print.sgdnet]{print.sgdnet()}}
}
