% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prostate}
\alias{prostate}
\title{prostate}
\format{A dataframe with two items representing 97 observations from
8 variables
\describe{
\item{lcavol}{log(cancer volume)}
\item{lweight}{log(prostate weight)}
\item{age}{age}
\item{lbph}{log(benign prostatic hyperplasia amount)}
\item{svi}{seminal vesicle invasion}
\item{lcp}{log(capsular penetration)}
\item{gleason}{weight of meat}
\item{pgg45}{Gleason score}
}}
\source{
Stamey, T.A., Kabalin, J.N., McNeal, J.E., Johnstone, I.M., Freiha, F., Redwine, E.A. and Yang, N. (1989),
Prostate specific antigen in the diagnosis and treatment of adenocarcinoma of the prostate: II. radical prostatectomy treated patients, Journal of Urology 141(5), 1076–1083.
}
\usage{
prostate
}
\description{
This data come from a study that examined the correlation
between the level of prostate specific antigen
and a number of clinical measures in men who were about to receive a radical prostatectomy.
}
\keyword{datasets}
