% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.sgdnet.R
\name{refit}
\alias{refit}
\title{Refit sgdnet object}
\usage{
refit(object, s, ...)
}
\arguments{
\item{object}{an object of class \code{sgdnet}}

\item{s}{a new lambda penalty (or several)}

\item{...}{new arguments to pass on to \code{\link[=sgdnet]{sgdnet()}}}
}
\value{
A new fit from \code{\link[=sgdnet]{sgdnet()}}.
}
\description{
Refit sgdnet object
}
\keyword{internal}
