% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\alias{score.sgdnet_gaussian}
\alias{score.sgdnet_binomial}
\alias{score.sgdnet_multinomial}
\alias{score.sgdnet_mgaussian}
\alias{score.sgdnet_poisson}
\alias{score.cv_sgdnet}
\title{Score the Performance of a Model fit with \code{\link[=sgdnet]{sgdnet()}}}
\usage{
score(fit, ...)

\method{score}{sgdnet_gaussian}(fit, x, y, type.measure = c("deviance",
  "mse", "mae"), s = fit$lambda, ...)

\method{score}{sgdnet_binomial}(fit, x, y, type.measure = c("deviance",
  "mse", "mae", "class", "auc"), s = fit$lambda, ...)

\method{score}{sgdnet_multinomial}(fit, x, y,
  type.measure = c("deviance", "mse", "mae", "class"), s = fit$lambda,
  ...)

\method{score}{sgdnet_mgaussian}(fit, x, y, type.measure = c("deviance",
  "mse", "mae"), s = fit$lambda, ...)

\method{score}{sgdnet_poisson}(fit, x, y, type.measure = c("deviance",
  "mse", "mae"), s = fit$lambda, ...)

\method{score}{cv_sgdnet}(fit, x, y, type.measure, s = c("lambda_1se",
  "lambda_min"), ...)
}
\arguments{
\item{fit}{the model fit}

\item{...}{arguments passed on to \code{\link[=predict.sgdnet]{predict.sgdnet()}}}

\item{x}{a feature matrix of new data}

\item{y}{response(s) for new data}

\item{type.measure}{the type of measure}

\item{s}{lambda}
}
\value{
Returns the prediction error along the lambda path.
}
\description{
The main purpose of this function is internal use within \code{\link[=cv_sgdnet]{cv_sgdnet()}},
where it is used to score the performance over folds in cross-validation.
It can, however, be used on its own to measure performance against a
validation set, for instance by training the model via \code{\link[=cv_sgdnet]{cv_sgdnet()}}
and holding out a validation set for
use with this function.
}
\examples{
set.seed(1)
n <- nrow(wine$x)
train_ind <- sample(n, floor(0.8*n))
cv_fit <- cv_sgdnet(wine$x[train_ind, ],
                    wine$y[train_ind],
                    family = "multinomial",
                    nfolds = 5,
                    alpha = c(0.5, 1))
score(cv_fit, wine$x[-train_ind, ], wine$y[-train_ind], "deviance")
}
\seealso{
\code{\link[=predict.sgdnet]{predict.sgdnet()}}, \code{\link[=cv_sgdnet]{cv_sgdnet()}}
}
