% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdnet-package.R
\docType{package}
\name{sgdnet-package}
\alias{sgdnet-package}
\title{Penalized Generalized Linear Models with Stochastic Gradient Descent}
\description{
Penalized Generalized Linear Models with Stochastic Gradient Descent
}
\section{Package options}{


Parts of the \link{sgdnet} API can be interfaced with via
\code{\link{options}} as follows:

\itemize{
\item \code{sgdnet.debug}: set to \code{TRUE} to enable debugging features,
which will be accessible via a \code{'diagnostics'}
attribute to the \code{'sgdnet'} object
that is returned by \code{\link[=sgdnet]{sgdnet()}}.
}
}

