% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdnet.R
\name{sgdnet}
\alias{sgdnet}
\alias{sgdnet.default}
\title{Fit a Generalized Linear Model with Elasticnet Regularization}
\usage{
sgdnet(x, y, ...)

\method{sgdnet}{default}(x, y, family = c("gaussian"), alpha = 1,
  lambda = 1/NROW(x), maxit = 1000, standardize = TRUE,
  intercept = TRUE, thresh = 0.001, ...)
}
\arguments{
\item{x}{input matrix}

\item{y}{response variable}

\item{...}{ignored}

\item{family}{reponse type}

\item{alpha}{elasticnet mixing parameter}

\item{lambda}{regularization strength}

\item{maxit}{maximum number of effective passes (epochs)}

\item{standardize}{whether to standardize \code{x} or not -- ignored when
\code{intercept == TRUE}.}

\item{intercept}{whether to fit an intercept or not}

\item{thresh}{tolerance level for termination of the algorithm}
}
\value{
An object of class \code{'sgdnet'}.
}
\description{
Fit a Generalized Linear Model with Elasticnet Regularization
}
\examples{
# Gaussian regression with sparse features
sgdnet(permeability$x, permeability$y, alpha = 0)
}
