% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SgdnetCpp}
\alias{SgdnetCpp}
\title{Fit a Model with sgdnet}
\usage{
SgdnetCpp(x_in, y, control)
}
\arguments{
\item{x_in}{feature matrix}

\item{y}{response matrix}

\item{control}{a list of control parameters}
}
\value{
A list of
\itemize{
\item ao: the intercept,
\item beta: the weights,
\item losses: the loss at each outer iteration per fit,
\item npasses: the number of effective passes (epochs) accumulated over,
all lambda values, and
\item return_codes: the convergence result. 0 mean that the algorithm
converged, 1 means that \code{max_iter} was reached before the algorithm
converged.
}
}
\description{
This main use of this function is calling the templated SetupSgdnet()
so that the dense and sparse implementations are compiled and
called appropriately. The control parameters in \code{control} are just
passed along.
}
\keyword{internal}
