% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.sgdnet.R
\name{softmax}
\alias{softmax}
\title{Softmax-based prediction}
\usage{
softmax(x)
}
\arguments{
\item{x}{linear predictors}
}
\value{
Class predictions.
}
\description{
Softmax-based prediction
}
\keyword{internal}
