% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdnet_methods.R
\name{coef.sgdnet}
\alias{coef.sgdnet}
\title{Extract Model Coefficients for sgdnet Model}
\usage{
\method{coef}{sgdnet}(object, ...)
}
\arguments{
\item{object}{a model of class \code{'sgdnet'}, typically from a call to
\code{\link[=sgdnet]{sgdnet()}}.}

\item{...}{ignored}
}
\value{
A sparse matrix with intercept in the first row and betas
in the rest.
}
\description{
Extract Model Coefficients for sgdnet Model
}
\examples{
fit <- sgdnet(matrix(rnorm(100), 50, 2), rnorm(50))
coef(fit)
}
