% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sgdnet.R
\name{plot.sgdnet}
\alias{plot.sgdnet}
\title{Plot Coefficients from an 'sgdnet' Object}
\usage{
\method{plot}{sgdnet}(x, xvar = c("norm", "lambda", "dev"), ...)
}
\arguments{
\item{x}{an object of class \code{'sgdnet'}, commonly the result from calling
\code{\link[=sgdnet]{sgdnet()}}.}

\item{xvar}{value to be plotted on the x axis. \code{"norm"} plots the
L1 norm, \code{"lambda"} the logarithmized lambda (penalty) values, and \code{"dev"}
the percent of deviance explained.}

\item{...}{parameters passed down to \code{\link[lattice:xyplot]{lattice::xyplot()}}.}
}
\value{
A graphical description of class \code{'trellis'}, which will be
plotted on the current graphical device in interactive sessions.
}
\description{
Plot coefficients from an object of class \code{'sgdnet'} against the L1-norm,
lambda penalty, or deviance ratio.
}
\details{
This function calls \code{\link[lattice:xyplot]{lattice::xyplot()}} under the hood after having
arranged the plotting data slightly.
}
\examples{
# Gaussian logistic regression
gfit <- sgdnet(abalone$x, abalone$y, alpha = 0)
plot(gfit, auto.key = list(columns = 2, space = "top"))

# Binomial logistic regression
bfit <- sgdnet(with(infert, cbind(age, parity)),
               infert$case,
               family = "binomial")
plot(bfit, xvar = "lambda", grid = TRUE)

# Multinomial logistic regression
mfit <- sgdnet(iris[, 1:4], iris[, 5], family = "multinomial")
plot(mfit, xvar = "dev", main = "Lassoing with sgdnet")

# Multivariate gaussian regression
mgfit <- sgdnet(student$x, student$y, family = "mgaussian")
plot(mgfit, lty = 1:9, layout = c(1, 2))
}
\seealso{
\code{\link[lattice:xyplot]{lattice::xyplot()}}, \code{\link[=sgdnet]{sgdnet()}}
}
