% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdnet.R
\name{sgdnet}
\alias{sgdnet}
\alias{sgdnet.default}
\title{Fit a Generalized Linear Model with Elastic Net Regularization}
\usage{
sgdnet(x, ...)

\method{sgdnet}{default}(x, y, family = c("gaussian", "binomial"),
  alpha = 1, nlambda = 100, lambda.min.ratio = if (NROW(x) < NCOL(x)) 0.01
  else 1e-04, lambda = NULL, maxit = 1000, standardize = TRUE,
  intercept = TRUE, thresh = 0.001, ...)
}
\arguments{
\item{x}{input matrix}

\item{...}{ignored}

\item{y}{response variable}

\item{family}{reponse type, one of \code{'gaussian'} and \code{'binomial'}. See
\strong{Supported families} for details.}

\item{alpha}{elastic net mixing parameter}

\item{nlambda}{number of penalties in the regualrization path}

\item{lambda.min.ratio}{the ratio between \code{lambda_max} (the smallest
penalty at which the solution is completely sparse) and the smallest
lambda value on the path. See section \strong{Regularization Path} for details.}

\item{lambda}{regularization strength}

\item{maxit}{maximum number of effective passes (epochs)}

\item{standardize}{whether to standardize \code{x} or not -- ignored when
\code{intercept == FALSE}.}

\item{intercept}{whether to fit an intercept or not}

\item{thresh}{tolerance level for termination of the algorithm. The
algorithm terminates when
\deqn{
    \frac{|\beta^{(t)}
    - \beta^{(t-1)}|{\infty}}{|\beta^{(t)}|{\infty}} < \mathrm{thresh}
  }{
    max(change in weights)/max(weights) < thresh.
  }}
}
\value{
An object of class \code{'sgdnet'}.
}
\description{
Fit a Generalized Linear Model with Elastic Net Regularization
}
\section{Model families}{

Two model families are currently supported: gaussian univariate
regression and binomial logistic regression, the choice of which is made
using the \code{family} argument.

The \emph{gaussian} family solves the following objective:

\deqn{
  \frac{1}{2n} \sum_{i=1}^n (y_i - \beta_0 - x_i^\mathsf{T} \beta)^2
  + \lambda \left( \frac{1 - \alpha}{2} ||\beta||_2^2
                   + \alpha||\beta||_1 \right).
}{
  1/(2n) \sum (y - \beta_0 - x^T \beta)^2
  + \lambda [(1 - \alpha)/2 ||\beta||_2^2 + \alpha||\beta||_1].
}

The \emph{binomial} family solves the following objective:

\deqn{
  -\frac{1}{n} \sum_{i=1}^n
    \bigg[y_i (\beta_0 + x_i^\mathsf{T} \beta) - \log\Big(1 + e^{\beta_0 + x_i^\mathsf{T} \beta}\Big)\bigg]
  + \lambda \left( \frac{1 - \alpha}{2} ||\beta||_2^2
                   + \alpha||\beta||_1 \right),
}{
  -1/n \sum [y_i(\beta_0 + x^T \beta) - log(1 + exp(\beta_0 + x^T \beta)]
  + \lambda [(1 - \alpha)/2 ||\beta||_2^2 + \alpha||\beta||_1],
}
where \eqn{y_i \in \{0, 1\}}{y ~ {0, 1}}.
}

\section{Regularization Path}{

The default regularization path is a sequence of \code{nlambda}
log-spaced elements
from \eqn{\lambda_{\mathrm{max}}}{lambda_max} to
\eqn{\lambda_{\mathrm{max}} \times \mathtt{lambda.min.ratio}}{
     lambda_max*lambda.min.ratio},
For the gaussian family, for instance,
\eqn{\lambda_{\mathrm{max}}}{lambda_max} is
the largest absolute inner product of the feature vectors and the response
vector,
\deqn{\max_i \frac{1}{n}|\langle\mathbf{x}_i, y\rangle|.}{
      max |<x, y>|/n.}
}

\examples{
# Gaussian regression with sparse features
fit <- sgdnet(permeability$x, permeability$y, alpha = 0)
}
\seealso{
\code{\link[=predict.sgdnet]{predict.sgdnet()}}, \code{\link[=plot.sgdnet]{plot.sgdnet()}}, \code{\link[=coef.sgdnet]{coef.sgdnet()}},
\code{\link[=sgdnet-package]{sgdnet-package()}}
}
