% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdnet.R
\name{sgdnet}
\alias{sgdnet}
\alias{sgdnet.default}
\title{Fit a Generalized Linear Model with Elastic Net Regularization}
\usage{
sgdnet(x, y, ...)

\method{sgdnet}{default}(x, y, family = c("gaussian"), alpha = 1,
  lambda = 1/NROW(x), maxit = 1000, standardize = TRUE,
  intercept = TRUE, thresh = 0.001, return_loss = FALSE, ...)
}
\arguments{
\item{x}{input matrix}

\item{y}{response variable}

\item{...}{ignored}

\item{family}{reponse type}

\item{alpha}{elasticnet mixing parameter}

\item{lambda}{regularization strength}

\item{maxit}{maximum number of effective passes (epochs)}

\item{standardize}{whether to standardize \code{x} or not -- ignored when
\code{intercept == TRUE}.}

\item{intercept}{whether to fit an intercept or not}

\item{thresh}{tolerance level for termination of the algorithm}

\item{return_loss}{whether to compute and return the loss at each outer
iteration, only added here for debugging purposes.}
}
\value{
An object of class \code{'sgdnet'}.
}
\description{
Fit a Generalized Linear Model with Elastic Net Regularization
}
\examples{
# Gaussian regression with sparse features
x <- Matrix::rsparsematrix(100, 10, density = 0.2)
y <- rnorm(100)
sgdnet(x, y)
}
