% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdnet.R
\name{sgdnet}
\alias{sgdnet}
\alias{sgdnet.default}
\title{Fit a Generalized Linear Model with Elastic Net Regularization}
\usage{
sgdnet(x, y, ...)

\method{sgdnet}{default}(x, y, family = c("gaussian"), alpha = 1,
  nlambda = 100, lambda.min.ratio = if (NROW(x) < NCOL(x)) 0.01 else 1e-04,
  lambda = NULL, maxit = 1000, standardize = TRUE, intercept = TRUE,
  thresh = 0.001, ...)
}
\arguments{
\item{x}{input matrix}

\item{y}{response variable}

\item{...}{ignored}

\item{family}{reponse type}

\item{alpha}{elastic net mixing parameter}

\item{lambda}{regularization strength}

\item{maxit}{maximum number of effective passes (epochs)}

\item{standardize}{whether to standardize \code{x} or not -- ignored when
\code{intercept == FALSE}.}

\item{intercept}{whether to fit an intercept or not}

\item{thresh}{tolerance level for termination of the algorithm. The
algorithm terminates when
\deqn{
    \frac{|\beta^{(t)} - \beta^{(t-1)}|{\infty}}{|\beta^{(t)}|{\infty}} < \mathrm{thresh}
  }{
    max(change in weights)/max(weights) < thresh.
  }}
}
\value{
An object of class \code{'sgdnet'}.
}
\description{
Fit a Generalized Linear Model with Elastic Net Regularization
}
\examples{
# Gaussian regression with sparse features
fit <- sgdnet(permeability$x, permeability$y, alpha = 0)
}
